fixed GPSHandler... to test with payload code

+++ lib/GPSHandler/GPSHandler.h
@@ -10,8 +10,8 @@
  * Date Created: 9 Mar 2019
  * Description: A custom library to help in parsing raw GPS data
  * Last Edited By: Jordan Martin
- * Date Edited: 11 Mar 2019
- * Reason Edited: Added getRawData()
+ * Date Edited: 25 Mar 2019
+ * Reason Edited: Fixed DMtoDecimal (wasn't of type float)
  */
 #define GPS_FIXED 1
 #define GPS_NOT_FIXED 0
@@ -34,7 +34,7 @@ public:
   int getRawData(String *data); // gets raw gps data
 private:
   int parseData(String toParse, GPSData *parsedData); // parses gps data
-  int DMtoDecimal(String dm); // converts gps lat and lng from DDMM.MMMM (degMin.Min) to degrees
+  float DMtoDecimal(String dm); // converts gps lat and lng from DDMM.MMMM (degMin.Min) to degrees
   bool dataGood = false;
   long timeAtLastUpdate = 0;
   BetterSerial gpsSerial;
